% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_gene_cor.R
\name{all_gene_cor}
\alias{all_gene_cor}
\title{Computes gene-gene correlations}
\usage{
all_gene_cor(ge, cell.size = NULL, return.unwt.cor = FALSE)
}
\arguments{
\item{ge}{super-cell gene expression matrix}

\item{cell.size}{super-cell size}
}
\value{
list with fields:
 \code{cell.size} - super-cell size, \code{cor} unweighted correlation (\code{if(return.unwt.cor)}), \code{wt.cor} - weighted correlation betweeen genes, \code{wt.pval} p.values of weighted correlations
}
\description{
Computes gene-gene correlations
}
