% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_supercells_to_clusters.R
\name{annotate_supercells_to_cluster}
\alias{annotate_supercells_to_cluster}
\title{Computes annotation of super-cell types based on the annotation of single cells and purity of super-cell}
\usage{
annotate_supercells_to_cluster(
  SC.list,
  sc.annotation,
  annotation.name,
  annotation.meth = "jaccard",
  verbose = FALSE
)
}
\arguments{
\item{SC.list}{list of super-cell-like structures (output of \code{compute_supercells})}

\item{sc.annotation}{vector of cell type / cluster / etc annotation at the single-cell level (length on the vector == number of single cells)}

\item{annotation.name}{a name of the annotation, will be added as additional SC.list field}

\item{annotation.meth}{\code{method} parameter of \link[SuperCell]{supercell_assign} function. Default is 'jaccard'}

\item{verbose}{flag whether to pring progress steps}
}
\value{
SC.list with each super-cell element containing additional field \code{annotation.name} that contains super-cell annotation
}
\description{
Runs \code{\link[SuperCell]{supercell_assign}} and \code{\link[SuperCell]{supercell_purity}} for all super-cell elements in SC.list (output of \code{compute_supercells})
}
