% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_supercell_like_structure.R
\name{bind_by_graining_level}
\alias{bind_by_graining_level}
\title{Concatenates super-cell-like structures by graining level}
\usage{
bind_by_graining_level(SC.list1, SC.list2)
}
\arguments{
\item{SC.list2}{}
}
\value{
concatenated `SC.list`
}
\description{
Concatenates super-cell-like structures by graining level
}
