% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_supercell_like_structure.R
\name{bind_by_method}
\alias{bind_by_method}
\title{Concatenates super-cell-like structures by method}
\usage{
bind_by_method(SC.list1, SC.list2)
}
\value{
concatenated `SC.list`
}
\description{
Concatenates super-cell-like structures by method
}
