% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_MC.R
\name{build_MC}
\alias{build_MC}
\title{Build metacell oblect
Wrapper that inits and builds metacell object from count matrix}
\usage{
build_MC(
  sc.counts,
  proj.name,
  genes.to.use = NULL,
  mc.k.knn = 100,
  MC.T_vm = 0.08,
  MC.folder = "MC",
  min_mc_size = 20,
  p_resamp = 0.75,
  n_resamp = 500,
  mc.alpha = 2,
  mc.K = 30,
  verbose = FALSE
)
}
\arguments{
\item{sc.counts}{single-cell count matrix (genes x cells)}

\item{proj.name}{project_name parameter in \link[metacell]}

\item{genes.to.use}{genes to use ti build metacell}

\item{mc.k.knn}{k in kNN single-cell network in metacell}

\item{MC.T_vm}{cutoff to select HVG in metacell}

\item{MC.folder}{folder to store metacell files}

\item{min_mc_size}{minimum size of metacell (used to vary graining level)}

\item{p_resamp}{}
}
\value{
metacell::scdb_mc - like output with some additional information stored as a list
}
\description{
Build metacell oblect
Wrapper that inits and builds metacell object from count matrix
}
