% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_supercells_DEA.R,
%   R/compute_supercells_clustering.R
\name{compute_alternative_clustering}
\alias{compute_alternative_clustering}
\title{Compute alternative clustering at single-cell level}
\usage{
compute_alternative_clustering(
  sc.pca,
  N.comp = 10,
  N.clusters.seq = c(2:10),
  hclust_methods = c("ward.D", "ward.D2", "single", "complete", "average", "mcquitty",
    "median", "centroid"),
  other_clust_funcs = list(kmeans = kmeans),
  seed.seq = c(12346, 111, 19, 42, 7)
)

compute_alternative_clustering(
  sc.pca,
  N.comp = 10,
  N.clusters.seq = c(2:10),
  hclust_methods = c("ward.D", "ward.D2", "single", "complete", "average", "mcquitty",
    "median", "centroid"),
  other_clust_funcs = list(kmeans = kmeans),
  seed.seq = c(12346, 111, 19, 42, 7)
)
}
\arguments{
\item{sc.pca}{PCA matrix or high dimensional matrix (cells as rows and coordinates/genes as columns)}

\item{N.comp}{number or vector of PCs to use}

\item{N.clusters.seq}{vector of number of clusters to compute}

\item{hclust_methods}{a vector of \code{method} parameter form \link[stats]{hclust}}

\item{other_clust_funcs}{list of ther clustering functions (in a format of \link[stats]{kmeans})}
}
\value{
list of clusterings \code{res} in a format \code{res[["number_of_clusters"]][["clustering_method[_random_seed]"]]}

list of clusterings \code{res} in a format \code{res[["number_of_clusters"]][["clustering_method[_random_seed]"]]}
}
\description{
Compute alternative clustering at single-cell level

Compute alternative clustering at single-cell level
}
