% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_alternative_SC.R
\name{compute_alternative_SC}
\alias{compute_alternative_SC}
\title{Compute super-cells using alternative algorithms behind.
Compute super-cells trying different number of neighbors for KNN, SNN implements in Seurat and also different clustering algorithms (i.e., current walktrap and Seurat-based Louvain)}
\usage{
compute_alternative_SC(
  counts,
  genes.use,
  meta.data = NULL,
  ge = NULL,
  N.comp = 10,
  k.seq = sort(c(5, 10, 50, 100, round(0.01 * ncol(counts)), round(0.05 *
    ncol(counts)))),
  res.seq = c(1, 2, 5, 10, 20, 30, 40, 50, 60, 70, 80),
  gamma.seq = NULL,
  return.singlecell.NW = FALSE,
  verbose = FALSE,
  common.gammas = TRUE,
  do.directed = c(T, F),
  group.singletons = FALSE,
  max.gamma = 150,
  ...
)
}
\arguments{
\item{counts}{count matrix (genes x cells)}

\item{genes.use}{a set of genes used in the original super-cell construction (output of \link[SuperCell]{SCimplify}, field `genes.use`)}

\item{meta.data}{meta data as an input to \link[SeuratObject]{CreateSeuratObject}}

\item{ge}{log-normalized gene expression matrix (if computation is different from whar Seurat outputs)}

\item{k.seq}{set of k to compute knn network}

\item{res.seq}{set of resolutions to obtain different graining levels using louvain clustering in \link[Seurat]{FindClusters}}

\item{gamma.seq}{graining levels to compute for walktrap clustering (is NULL, graining levels are retreived from the graining levels obtained in louvain clustering with \code{res.seq} resolutions)}

\item{return.singlecell.NW}{whether to return single-cell network}

\item{common.gammas}{whether to compute walktrap super-cells for both provided graining leveles (if providede) and obtained graning levels with louvain}
}
\value{
list of SuperCell -like structures with the first result layer corresponding to a NW type and the second layer corresponding to the clustering type (walktrap and louvain for the moment), the third layer is a graining level
}
\description{
Compute super-cells using alternative algorithms behind.
Compute super-cells trying different number of neighbors for KNN, SNN implements in Seurat and also different clustering algorithms (i.e., current walktrap and Seurat-based Louvain)
}
