% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_supercells_clustering.R
\name{compute_consistency_of_supercell_clustering}
\alias{compute_consistency_of_supercell_clustering}
\title{Compute consistency of super-cell (meta-cell) clustering with the GT annotation (or single-cell clustering, if GT annotation is not available)}
\usage{
compute_consistency_of_supercell_clustering(
  SC.list,
  sc.annotation,
  sc.clustering = NULL,
  clustering.name = "hclust",
  sc.alternative.clustering = NULL,
  verbose = FALSE
)
}
\arguments{
\item{SC.list}{list of super-cells and other simplifications (output of \link{compute_supercell})}

\item{sc.clustering}{clustering of single-cell data, if not provided (is NULL), will be set to \code{GT_annotation}}

\item{sc.alternative.clustering}{output of \link{compute_alternative_clustering} or NULL}

\item{GT_annotation}{ground truth clustering of single-cells}

\item{clustering_name}{names of clustering (default is 'hclust')}
}
\description{
Compute consistency of super-cell (meta-cell) clustering with the GT annotation (or single-cell clustering, if GT annotation is not available)
}
