% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_supercells_DEA.R
\name{compute_consistency_of_supercell_DEA}
\alias{compute_consistency_of_supercell_DEA}
\title{Compute consistency of super-cell (meta-cell) DEA with the GT DEA (DEA for GT annotation)}
\usage{
compute_consistency_of_supercell_DEA(
  DEA.list,
  GT.DEA.bool,
  sc.DEA = NULL,
  verbose = FALSE
)
}
\arguments{
\item{DEA.list}{list of DEA result (output of \link{compute_supercells_DEA})}

\item{GT.DEA.bool}{ground truth differentially expressed genes (named boolean with names beeing genes)}

\item{sc.DEA}{if NULL (dafault) equalt to \code{GT.DEA.bool}}
}
\value{
TPR, F1, AUC between GT and super-cell DEA
}
\description{
Compute consistency of super-cell (meta-cell) DEA with the GT DEA (DEA for GT annotation)
}
