% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_supercells_DEA.R
\name{compute_DEA_GT_bool}
\alias{compute_DEA_GT_bool}
\title{Compute bool vector of cluster-specific DEG}
\usage{
compute_DEA_GT_bool(
  DEA.GT,
  all.genes = NULL,
  logFC.thresh = 1,
  pval.thresh = 0.05
)
}
\arguments{
\item{DEA.GT}{GT differentially expressed genes (DEA for GT annotation)}

\item{logFC.thresh}{thereshold to consider DEG to be statistically significant}

\item{pval.thresh}{thereshold to consider DEG to be statistically significant (keep the same as in \link{compute_supercells_DEA} and \link{compute_singlecells_DEA})}
}
\description{
Compute bool vector of cluster-specific DEG
}
