% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_singlecell_silh.R
\name{compute_singlecell_silh}
\alias{compute_singlecell_silh}
\title{Computes silhouette width for a sequence of single-cell clustering}
\usage{
compute_singlecell_silh(clustering.list, d)
}
\arguments{
\item{clustering.list}{list of clustering result with names of the field being number of clusters and value being clusterting result}

\item{d}{cell-to-cell distance (dissimilarity)}
}
\value{
names vector with names being number of clusters and values being average silhouette coefficient
}
\description{
Computes silhouette width for a sequence of single-cell clustering
}
