% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_supercells_clustering.R
\name{compute_singlecell_unweighted_clustering}
\alias{compute_singlecell_unweighted_clustering}
\title{Coumputes unweighted clustering for single-cell data}
\usage{
compute_singlecell_unweighted_clustering(
  sc.pca,
  sc.GE,
  N.clusters.seq,
  sc.dist = NULL,
  N.clusters = NULL,
  genes.use = NULL,
  algs = c("hclust", "kmeans", "seurat"),
  N.comp = NULL,
  DO_silhouette = TRUE,
  seurat_fields = c(),
  def.resolution = 0.8,
  step.resolution = 0.1,
  max.counter = 20,
  seed = 12345,
  verbose = FALSE
)
}
\arguments{
\item{sc.pca}{single-cell pca}
}
\description{
Coumputes unweighted clustering for single-cell data
}
