% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_supercells.R
\name{compute_supercells}
\alias{compute_supercells}
\title{Computes super-cells for the Exact, Approximate super-cells and for Subsampling and Random grouping}
\usage{
compute_supercells(
  sc.GE,
  gamma.seq,
  ToComputeSC,
  filename = "initial",
  data.folder = "./data",
  n.var.genes = 1000,
  k.knn = 5,
  n.pc = 10,
  approx.N = 1000,
  fast.pca = TRUE,
  genes.use = NULL,
  genes.exclude = NULL,
  seed.seq = c(12345, 111, 19, 42, 7)
)
}
\arguments{
\item{sc.GE}{single-cell gene expression matrix with genes as rows and cells as columns}

\item{gamma.seq}{a vector of graining levels (i.e., \code{c(1, 5, 10, 50, 100, 200)})}

\item{ToComputeSC}{a bool flag to indicate whether to compute super-cells (if TRUE) or load (if FALSE) (if was previously computed and saved)}

\item{filename}{filename to save/load the resulting list of super-cells}

\item{data.folder}{path to the data folder of a certain analysis, usualy in a format './data/project_name/'}

\item{n.var.genes}{\code{"n.var.genes"} parameter of SuperCell function \code{\link[SuperCell]{SCimplify()}}}

\item{k.knn}{\code{"k.knn"} parameter of SuperCell function \code{\link[SuperCell]{SCimplify()}}}

\item{n.pc}{\code{"n.pc"} parameter of SuperCell function \code{\link[SuperCell]{SCimplify()}}, can be a number, then 1:n.pc PC will be used to compute super-cells or a specific vector of PC to use (e.g., 2:15, c(2, 5, 6, 19))}

\item{approx.N}{\code{"approx.N"} parameter of SuperCell function \code{\link[SuperCell]{SCimplify()}}}

\item{fast.pca}{whethre to use fast PCA implementation from ilbra package (default TRUE)}

\item{genes.use}{\code{"genes.use"} parameter of SuperCell function \code{\link[SuperCell]{SCimplify()}}. A specific genes set to use for computing super-cells, if provided, then parameter \code{"n.var.genes"} will be omitted, if not provided, top \code{"n.var.genes"} most variable genes will be used}

\item{genes.exclude}{genes exclude from the gene set used for the computation of super-cells (e.g., a vector containing mito genes or ribo genes)}

\item{seed.seq}{a vector of Random seeds (used for the Approximate coarse-graining, Subsampling and Random grouping)}
}
\value{
list of super-cell in a format SC.list[[SC_method]][[gamma_i]][[seed_i]]
}
\description{
Computes super-cells for the Exact, Approximate super-cells and for Subsampling and Random grouping
for gammas in \code{"gamma.seq"} and for different Random seeds \code{"seed.seq"}
}
\examples{
\dontrun{

}
}
