% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_supercells.R
\name{compute_supercells_additional_seeds}
\alias{compute_supercells_additional_seeds}
\title{Computed super-cells for the additional seeds and concatenates the results to the existing ones}
\usage{
compute_supercells_additional_seeds(
  SC.list,
  additional_seed_seq,
  ToComputeSC,
  data.folder = "./data",
  filename = "additional_seeds",
  approx.N = 1000,
  fast.pca = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{SC.list}{output of \link{compute_supercells} (i.e., list of super-cells in a format \code{"SC.list[['SC_method']][[gamma_i]][[seed_i]]"})}

\item{additional_seed_seq}{vector of aggitional seeds for which SC has to be computed}
}
\value{
apdated list of super-cells with additional seeds
}
\description{
The same parameters as in \link{compute_supercells}, with 2 exceptions
}
