% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_supercells_DEA.R,
%   R/compute_supercells_clustering.R
\name{compute_supercells_clustering}
\alias{compute_supercells_clustering}
\title{Computes clustering for super-cells}
\usage{
compute_supercells_clustering(
  SC.list,
  N.comp = 10,
  N.clusters.seq = c(2:10),
  pca_name = "SC_PCA"
)

compute_supercells_clustering(
  SC.list,
  N.comp = 10,
  N.clusters.seq = c(2:10),
  pca_name = "SC_PCA"
)
}
\arguments{
\item{SC.list}{list of super-cells and other simplifications (output of \link{compute_supercell})}

\item{N.comp}{number or a vector of principal components to compute clastering on}

\item{N.clusters.seq}{vector of number of clusters to compute}

\item{pca_name}{name of the PCA field, default is 'SC_PCA' as returned with \link{compute_supercells_PCA}}
}
\value{
\code{SC.list} with additional fields:
 \code{hclust} with hierarchical clustering results and
 \code{silh:hclust} with silhouette coefficients

\code{SC.list} with additional fields:
 \code{hclust} with hierarchical clustering results and
 \code{silh:hclust} with silhouette coefficients
}
\description{
Computes clustering for super-cells

Computes clustering for super-cells
}
