% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_supercells_DEA.R
\name{compute_supercells_DEA}
\alias{compute_supercells_DEA}
\title{Computes differential expression analysis for super-cells}
\usage{
compute_supercells_DEA(
  SC.list,
  SC.GE.list,
  cluster.name,
  ident.1 = NULL,
  ident.2 = NULL,
  seed = 12345,
  verbose = FALSE,
  pval.thresh = 0.05,
  logFC.thresh = 0,
  DO_unweighted = FALSE,
  ...
)
}
\arguments{
\item{SC.list}{list of super-cells and other simplifications (output of \link{compute_supercell})}

\item{SC.GE.list}{list of super-cell gene expressions}

\item{cluster.name}{name of clusterig result field}

\item{ident.1}{name(s) of cluster for which markers are computed, if \code{NULL} (default), markers for all clusters will be computed (running \link[SuperCell]{supercell_FindAllMarkers} instead of \link[SuperCell]{supercell_FindMarkers}}

\item{ident.2}{name(s) of clusters for comparison. If \code{NULL} (defauld), then all the other clusters used}

\item{seed}{sandom seed to compute logFC_AUC score}

\item{DO_unweighted}{whether to compute not sample-weighed DEA}

\item{...}{other parameters of \link[SuperCell]{supercell_FindMarkers} or \link[SuperCell]{supercell_FindAllMarkers}}
}
\value{
list of DEA results for each super-cell like structure
}
\description{
Computes differential expression analysis for super-cells
}
