% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_supercells_GE.R
\name{compute_supercells_GE}
\alias{compute_supercells_GE}
\title{Computes gene expression of metacells}
\usage{
compute_supercells_GE(
  sc.GE,
  SC.list,
  ToComputeSC_GE = TRUE,
  filename = "additional_gammas",
  data.folder = "./data",
  folder.suf = "",
  verbose = FALSE
)
}
\arguments{
\item{sc.GE}{single-cell gene expression}

\item{SC.list}{list of super-cells (an output of \code{\link{compute_supercells}} or \code{\link{compute_supercells_metacells}})}

\item{ToComputeSC_GE}{whether to compute or to load already computed GE}

\item{filename}{file name where to save or load GE}

\item{data.folder}{path to the data folder, usually has a fprmat './data/project_name'}
}
\value{
a list of GE in the same format as \code{"SC.list"}, i.e., \code{"SC.GE.list[[MS_method]][[gamma_i]][[seed_i]]"}
}
\description{
Computes gene expression of metacells
}
