% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_supercells_metacells.R
\name{compute_supercells_metacells}
\alias{compute_supercells_metacells}
\title{compute super-cell like structure of metacells}
\usage{
compute_supercells_metacells(
  sc.counts,
  gamma.seq,
  SC.list,
  proj.name,
  ToComputeSC,
  mc.k.knn = 100,
  T_vm_def = 0.08,
  MC.folder = "MC",
  MC_gene_settings = c("Metacell_default", "Metacell_SC_like")
)
}
\arguments{
\item{ToComputeSC}{if TRUE computes metacells, if FALSE, loads from corresponding folder}
}
\description{
compute super-cell like structure of metacells
}
