% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_supercells_metacells.R
\name{compute_supercells_metacells_with_min_mc_size}
\alias{compute_supercells_metacells_with_min_mc_size}
\title{compute super-cell like structure of metacells (based on provided `min_mc_size` instead of `min_mc_size` as min supercell size)}
\usage{
compute_supercells_metacells_with_min_mc_size(
  sc.counts,
  gamma.seq,
  SC.list,
  min_mc_size_seq,
  proj.name,
  ToComputeSC,
  mc.k.knn = 100,
  T_vm_def = 0.08,
  MC.folder = "MC",
  MC_gene_settings = c("Metacell_default", "Metacell_SC_like"),
  verbose = FALSE
)
}
\arguments{
\item{ToComputeSC}{if TRUE computes metacells, if FALSE, loads from corresponding folder}
}
\description{
compute super-cell like structure of metacells (based on provided `min_mc_size` instead of `min_mc_size` as min supercell size)
}
