% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_supercells_PCA.R
\name{compute_supercells_PCA}
\alias{compute_supercells_PCA}
\title{Computes PCA for supee-cells over all simplification methods, graining levels (gammas) and random seeds}
\usage{
compute_supercells_PCA(
  SC.list,
  SC.GE.list,
  N.comp,
  genes.omit = NULL,
  pca_name = "SC_PCA",
  verbose = FALSE
)
}
\arguments{
\item{SC.list}{super-cell list}

\item{SC.GE.list}{super-cell gene expression list}

\item{N.comp}{numbe of PC to compute}

\item{genes.omit}{list of genes to exclude from hvg}

\item{pca_name}{name of the fiels to add to SC.list}
}
\value{
\code{SC.list} with an additional fiels named \code{pca_name} containing PCA embedding
}
\description{
Computes PCA for supee-cells over all simplification methods, graining levels (gammas) and random seeds
}
