% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_supercells_clustering.R
\name{compute_supercells_silhouette}
\alias{compute_supercells_silhouette}
\title{Computes silhouete distance for a list of clustring results}
\usage{
compute_supercells_silhouette(
  SC.list,
  N.comp,
  pca_name = "SC_PCA",
  clustering_name = "hclust",
  verbose = FALSE
)
}
\arguments{
\item{SC.list}{super-cell-like structures}

\item{N.comp}{number of principal components to use for distance calculation}

\item{pca_name}{name (key) to PCA results}

\item{clustering_name}{name (key) to clustering results. Clustering result are expected to be stores as a list with list names being number of custers and value being clustering partition}
}
\value{
SC.list with a silhouette field paste0("silh:", clustering_name)
}
\description{
Computes silhouete distance for a list of clustring results
}
