% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_consistency_mtrx.R
\name{diag_consistency_mtrx}
\alias{diag_consistency_mtrx}
\title{Diagonalize two clustering consistency matricies to map clustering results}
\usage{
diag_consistency_mtrx(m)
}
\arguments{
\item{m}{matrix of two clustering consistency (\code{as.matrix(table(clustering, GT_annotation))})}
}
\description{
Diagonalize two clustering consistency matricies to map clustering results
}
