% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_supercells_metacells.R
\name{get_actual_gammas_metacell}
\alias{get_actual_gammas_metacell}
\title{returns list of actual gammas of metacells}
\usage{
get_actual_gammas_metacell(SC.mc)
}
\arguments{
\item{SC.mc}{the output of \code{\link{compute_supercells_metacells}}}
}
\value{
sorted list of rounded to int actuall gammas
}
\description{
returns list of actual gammas of metacells
}
