% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_GO_df.R
\name{get_GO}
\alias{get_GO}
\title{Retreives and saves GO data}
\usage{
get_GO(
  genes,
  martDS = "hsapiens_gene_ensembl",
  data.folder = "/data",
  go_domain = "all",
  DO.save = TRUE
)
}
\arguments{
\item{genes}{vector of genes for which get and save GO ids}

\item{martDS}{mart data set}

\item{data.folder}{folder to save GO dataset}

\item{DO.save}{whether to save GO dataset}
}
\value{
\link[biomaRt]{getBM} result for genes
}
\description{
Retreives and saves GO data
}
