% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SC.R
\name{get_SC_by_gamma}
\alias{get_SC_by_gamma}
\title{Returns super-cell-like structure at specific graining levels}
\usage{
get_SC_by_gamma(SC.list, gamma.seq)
}
\arguments{
\item{SC.list}{super-cell--like structure}

\item{gamma.seq}{vector of gammas to exptact}
}
\value{
`SC.list` at specific graining levels
}
\description{
Returns super-cell-like structure at specific graining levels
}
