% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO_match_top_cor_genes.R
\name{GO_match_summary}
\alias{GO_match_summary}
\title{Computes mean and median values of jaccard GO match score within N.top statistically significant top correlated gene pairs}
\usage{
GO_match_summary(
  GO_match_SC,
  GO_match_sc,
  N.top = 1000,
  pval.thresh = 0.05,
  go_domain = "all",
  seed = 12345,
  na_to_0 = TRUE
)
}
\arguments{
\item{GO_match_SC}{list of \link{GO_match_top_cor_genes} of super-cells}

\item{GO_match_sc}{list of \link{GO_match_top_cor_genes} of single cells}

\item{N.top}{number of statistically significant top cor genes pais}

\item{pval.thresh}{statistical significance}

\item{go_domain}{GO domain to consider}

\item{na_to_0}{whether NA GO match replace with 0}
}
\value{
data frame with mean and median values of jaccard GO match score within N.top statistically significant top correlated gene pairs
}
\description{
Computes mean and median values of jaccard GO match score within N.top statistically significant top correlated gene pairs
}
