% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO_match_top_cor_genes.R
\name{GO_match_top_cor_genes}
\alias{GO_match_top_cor_genes}
\title{computes jaccard index of GO annotation of a pair of genes, used for an assessment of gene similarity (my approach)}
\usage{
GO_match_top_cor_genes(
  g.g.rank,
  bm,
  go_domain = c("biological_process", "molecular_function", "cellular_component")
)
}
\arguments{
\item{g.g.rank}{output of \link{rank_gene_gene_cor} or \link{add_cor_pval}}

\item{go_domain}{atr which GO domain (out of 'biological_process','molecular_function','cellular_component') to use, default all}

\item{martDS}{\code{dataset} parameter of \code{useMart} funcrion of package "biomaRt"}
}
\value{
g.g.rank data frame with additional fields \code{jaccard} and \code{shared}, where:
- \code{jaccard} jaccard index of GO annotation of a pair of genes
- \code{shared} absolute number of shared GO ids between a pair of genes
}
\description{
computes jaccard index of GO annotation of a pair of genes, used for an assessment of gene similarity (my approach)
}
