% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_alternative_SC.R
\name{Graph_to_igraph}
\alias{Graph_to_igraph}
\title{Convert \list[SeuratObject]{Graph} to \list[igraph]{igraph}}
\usage{
Graph_to_igraph(graph, ...)
}
\arguments{
\item{graph}{\list[SeuratObject]{Graph}}
}
\value{
\list[igraph]{igraph}
}
\description{
Convert \list[SeuratObject]{Graph} to \list[igraph]{igraph}
}
