% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_mc.R
\name{init_mc}
\alias{init_mc}
\title{Initiate MC folders}
\usage{
init_mc(mc.dir = NULL)
}
\arguments{
\item{mc.dir}{folder for MC data and outputs}
}
\value{
void
}
\description{
This function initiate MetaCell folder by running \code{scdb_init()}, \code{scfigs_init()}, \code{set_param()}
}
\examples{
\dontrun{

}
}
