% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_consistency_mtrx.R
\name{map_clustering_to_cell_type}
\alias{map_clustering_to_cell_type}
\title{Map super-cell clustering results to GT cell type annotation}
\usage{
map_clustering_to_cell_type(SC.list, GT.cell.type, clust.name)
}
\arguments{
\item{SC.list}{list of super-cell like structures}

\item{GT.cell.type}{vector with GT cell type annotation to which clusering has to be mapped}

\item{clust.name}{name of clustering that has to be mapped to \code{clust.name} (name of the SC.list field that corresponds to clustering result)}
}
\value{
SC.list with a new field named "\code{paste0(clust.name, "_mapped_to_GT")}" that contained clustering result of \code{clust.name} mapped to \code{GT.cell.type}
}
\description{
Map super-cell clustering results to GT cell type annotation
}
