% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_build_cgraph.R
\name{mc_build_cgraph}
\alias{mc_build_cgraph}
\title{Builds cell graph object using Metacell package}
\usage{
mc_build_cgraph(
  ge.mtrx,
  k.knn,
  genes.to.use = NULL,
  T_vm = -Inf,
  project_name = "metacell",
  balanced.Knn = TRUE,
  return.igraph = TRUE
)
}
\description{
Builds cell object of Metacell package using UMI data. No gene filtering performed inside,
 so either provide \code{ge.mtrx} as a UMI matrix with row containing only genes of interest or specify which genes to use via \code{genes.to.use}
}
\details{

}
