% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_build_cgraph.R
\name{mc_build_mc}
\alias{mc_build_mc}
\title{Builds metacell partition using output of \code{mc_build_cgraph}}
\usage{
mc_build_mc(
  X,
  min_mc_size = 15,
  p_resamp = 0.75,
  n_resamp = 500,
  mc.K = 30,
  mc.alpha = 2
)
}
\arguments{
\item{X}{-- an output of \code{mc_build_cgraph}}
}
\description{
Builds metacell partition using output of \code{mc_build_cgraph}
}
