% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_supercells_to_clusters.R
\name{plot_annotation_purity}
\alias{plot_annotation_purity}
\title{Plots super-cell purity}
\usage{
plot_annotation_purity(
  SC.list,
  annotation.name,
  SC_meth_exclude = c("Metacell_default_av", "Metacell_SC_like_av", "Subsampling"),
  seed = 12345,
  to.save.plot = TRUE,
  to.save.plot.raw = FALSE,
  asp = 0.5,
  fig.folder = "./plots",
  .shapes = c(Exact = 1, Approx = 0, Subsampling = 2, Random = 3, Metacell_default_fp =
    4, Metacell_default_av = 8, Metacell_SC_like_fp = 4, Metacell_SC_like_av = 8),
  .colors = c(Exact = "darkred", Approx = "royalblue", Subsampling = "black", Random =
    "gray", Metacell_default_fp = "forestgreen", Metacell_default_av = "forestgreen",
    Metacell_SC_like_fp = "gold", Metacell_SC_like_av = "gold"),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{SC.list}{list of super-cell-like structures (output of \code{compute_supercells})}

\item{annotation.name}{name of the reference annotation}
}
\description{
Plots super-cell purity
}
