% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_optimal_number_of_clusters.R
\name{plot_argmax_silh}
\alias{plot_argmax_silh}
\title{Plots optimal number of clusters (as argmax of silhouette coefficient)}
\usage{
plot_argmax_silh(
  SC.list,
  silh.name = "silh:hclust",
  sc.silh = NULL,
  error_bars = c("extr", "quartiles", "sd")[1],
  to.save.plot = TRUE,
  to.save.plot.raw = FALSE,
  asp = 0.5,
  fig.folder = "./plots",
  fig.name = "",
  ignore.gammas = c(),
  ignore.methods = c(),
  N.clusters.seq.breaks = NULL,
  bar_width = 0.1,
  dodge_width = 0.12,
  err_bar_line_size = 0.2,
  .shapes = c(Exact = 1, Approx = 0, Subsampling = 2, Random = 3, Metacell_default_fp =
    4, Metacell_default_av = 8, Metacell_SC_like_fp = 4, Metacell_SC_like_av = 8,
    Alternative = 23),
  .colors = c(Exact = "darkred", Approx = "royalblue", Subsampling = "black", Random =
    "gray", Metacell_default_fp = "forestgreen", Metacell_default_av = "forestgreen",
    Metacell_SC_like_fp = "gold", Metacell_SC_like_av = "gold", Alternative = "darkblue"),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{SC.list}{list of super-cell-like structures}

\item{silh.name}{name of silhouette field}

\item{sc.silh}{a vector of single-cell silhouettes with names being number of clusters and values being silhouettes coefficients}
}
\value{
barplot of (median) argmax silhouettes for a certain gamma and SC approach and corresponding data frame
}
\description{
Plots optimal number of clusters (as argmax of silhouette coefficient)
}
