% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_optimal_number_of_clusters.R
\name{plot_argmax_silh_alt_clusterings}
\alias{plot_argmax_silh_alt_clusterings}
\title{Plots optimal number of clusters (as argmax of silhouette coefficient) for multple clustering algorithms}
\usage{
plot_argmax_silh_alt_clusterings(
  SC.list,
  silh.names = c("silh:hclust", "silh:unw_hclust", "silh:unw_kmeans",
    "silh:unw_seurat"),
  sc.silh = NULL,
  error_bars = c("extr", "quartiles", "sd")[1],
  to.save.plot = TRUE,
  to.save.plot.raw = FALSE,
  asp = 0.5,
  fig.folder = "./plots",
  fig.name = "",
  ignore.gammas = c(),
  ignore.methods = c(),
  N.clusters.seq.breaks = NULL,
  bar_width = 0.1,
  dodge_width = 0.12,
  err_bar_line_size = 0.2,
  .shapes = c(hclust = 1, unw_hclust = 2, unw_kmeans = 2, unw_seurat = 2),
  .colors = c(hclust = "darkred", unw_hclust = "#882255", unw_kmeans = "#117733",
    unw_seurat = "#8CC6EC"),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{SC.list}{list of super-cell-like structures}

\item{sc.silh}{a vector of single-cell silhouettes with names being number of clusters and values being silhouettes coefficients}

\item{silh.name}{name of silhouette field}
}
\value{
barplot of (median) argmax silhouettes for a certain gamma and SC approach and corresponding data frame
}
\description{
Plots optimal number of clusters (as argmax of silhouette coefficient) for multple clustering algorithms
}
