% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO_match_top_cor_genes.R
\name{plot_GO_match_score}
\alias{plot_GO_match_score}
\title{Plot (normalized) GO match score}
\usage{
plot_GO_match_score(
  GO_summary,
  score.name = "GO_jaccard_Mean",
  error_bars = c("extr", "quartiles", "sd")[1],
  do.normalize = TRUE,
  to.save.plot = TRUE,
  to.save.plot.raw = FALSE,
  asp = 0.5,
  fig.folder = "./plots",
  fig.name = "",
  ignore.gammas = c(),
  ignore.methods = c(),
  .shapes = c(Exact = 1, Approx = 0, Subsampling = 2, Random = 3, Metacell_default_fp =
    4, Metacell_default_av = 8, Metacell_SC_like_fp = 4, Metacell_SC_like_av = 8),
  .colors = c(Exact = "darkred", Approx = "royalblue", Subsampling = "black", Random =
    "gray", Metacell_default_fp = "forestgreen", Metacell_default_av = "forestgreen",
    Metacell_SC_like_fp = "gold", Metacell_SC_like_av = "gold"),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{GO_summary}{a field of the output of \link{GO_match_summary}}

\item{score.name}{name of the score (one of the colnames of \code{GO_summary})}

\item{error_bars}{error bar types fo multiple testing simplifications (i.e., Approx, Subsampling, Random grouping)}

\item{do.normalize}{whether to normalize GO match scores with a single-cell GO match score}
}
\value{
plots GO match score and returns dataset used for plotting
}
\description{
Plot (normalized) GO match score
}
