% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_legend_cell_type.R
\name{plot_legend_celltype}
\alias{plot_legend_celltype}
\title{Plots and saves a legend mapping cell types and colors}
\usage{
plot_legend_celltype(pal.GT, fig.folder)
}
\arguments{
\item{pal.GT}{a palette as a named vector with names being cell types and values being corresponding colors}

\item{fig.folder}{folder for figures, usually has a format './plots/project_name'}
}
\value{
void, saves a legend plot to file.path(fig.folder, "Legend_cell_type.pdf")
}
\description{
Plots and saves a legend mapping cell types and colors
}
