% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_supercell_dropout_rate.R,
%   R/plot_supercell_size_distribution.R
\name{plot_supercell_size_distribution}
\alias{plot_supercell_size_distribution}
\title{Plot super-cell size distribution}
\usage{
plot_supercell_size_distribution(
  SC.list,
  methods_to_plot = NULL,
  ignore.gammas = NULL,
  fig.folder = "./plots",
  fig.name = "",
  doxplot.width = 1,
  outlier.alpha = 0,
  ...
)

plot_supercell_size_distribution(
  SC.list,
  methods_to_plot = NULL,
  ignore.gammas = NULL,
  fig.folder = "./plots",
  fig.name = "",
  doxplot.width = 1,
  outlier.alpha = 0,
  ...
)
}
\arguments{
\item{SC.list}{list of super-cell objects (result of \list[SuperCell]{SCimplify})}

\item{methods_to_plot}{methods to plot}

\item{ignore.gammas}{graining levels not to plot}

\item{fig.name}{fig name to add to the standard fig name

plots and saves}

\item{SC.GE.list}{list of super-cell gene expression}
}
\value{
data frame with super-cell size distributions

data frame with super-cell size distributions
}
\description{
Plot super-cell size distribution
}
