% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_gene_gene_cor.R
\name{rank_gene_gene_cor}
\alias{rank_gene_gene_cor}
\title{rank gene-gene correlation}
\usage{
rank_gene_gene_cor(
  gene.gene.cor,
  N.top = NULL,
  p.val.thresh = 0.05,
  rank.absolute = FALSE
)
}
\arguments{
\item{gene.gene.cor}{gene-gene correlation matrix (square)}

\item{N.top}{A number of top correlate genes to rank (by default all pairs are ranked)}
}
\value{
top.cor.genes -- ranking of pairwise correlation.
}
\description{
rank gene-gene correlation
}
