% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_mc_fields.R
\name{remove_mc_fields}
\alias{remove_mc_fields}
\title{Remove MC field from SC.list object when not needed anymore (i.e., when super-cell GE has been computed)}
\usage{
remove_mc_fields(SC.list)
}
\arguments{
\item{SC.list}{list of super-cell like strucctures for which mc fields have to be removed}
}
\description{
Remove MC field from SC.list object when not needed anymore (i.e., when super-cell GE has been computed)
}
