% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_alternative_SC.R
\name{SC_from_membership}
\alias{SC_from_membership}
\title{Builds super-cell object from a single-cell NW and a membership vector}
\usage{
SC_from_membership(
  sc.nw,
  membership,
  gname,
  resolution = NULL,
  N.comp = NULL,
  genes.use = NULL,
  return.singlecell.NW = FALSE
)
}
\arguments{
\item{sc.nw}{igraph single-cell network}

\item{membership}{super-cell membership vector}

\item{gname}{graph name}

\item{resolution}{clustering resolution (for louvain clustering)}

\item{N.comp}{number (or vector) of principal components used to build super-cells}

\item{genes.use}{genes used to build super-cells}

\item{return.singlecell.NW}{whether to return single-cell network}
}
\value{
the same as \link[SuperCell]{SCimplify} with an additional field - `res` for the resolution usign in \link[Seurat]{FindClusters} to obtain super-cell membership
}
\description{
Builds super-cell object from a single-cell NW and a membership vector
}
