% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveplot.R
\name{SCBM_saveplot}
\alias{SCBM_saveplot}
\title{Saves plot for manuscript}
\usage{
SCBM_saveplot(
  p,
  folder,
  name,
  w = 3,
  h = 3,
  ext = "pdf",
  save.raw.ggplot = TRUE,
  do.log = TRUE,
  asp = 1,
  gamma_field = "Gamma",
  do_filter_x_breaks = TRUE
)
}
\arguments{
\item{p}{ggplot object}

\item{folder}{figure folder}

\item{name}{figure name}

\item{w}{widht}

\item{h}{height}

\item{ext}{file expension}

\item{save.raw.ggplot}{flag whether to save raw ggplot}

\item{do.log}{do log normalization}

\item{asp}{plot aspect ratio}
}
\description{
Saves plot for manuscript
}
