% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCimplify2metacell.R
\name{SCimple2Metacell}
\alias{SCimple2Metacell}
\title{Computes metacell from super-cells}
\usage{
SCimple2Metacell(
  X,
  gamma = 10,
  min_mc_size = NULL,
  p_resamp = 0.75,
  n_resamp = 500,
  mc.alpha = 2,
  mc.K = 30
)
}
\description{
Computes metacell from super-cells
}
