% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCimplify2random.R
\name{SCimple2Random}
\alias{SCimple2Random}
\title{from Scimplify result to random simplification}
\usage{
SCimple2Random(SC, gamma, seed = 12345)
}
\description{
from Scimplify result to random simplification
}
