% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCimplify2subsampling.R
\name{SCimple2Subsampling}
\alias{SCimple2Subsampling}
\title{rom Scimplify result to subsampling}
\usage{
SCimple2Subsampling(X, SC, gamma, seed = 12345)
}
\description{
rom Scimplify result to subsampling
}
